Initial implementation of smooth scrolling.
This smooths out the rate streamed content is presented to the user.
It enables a more complex backend to provide intial results quickly
and then use the fact the user will be reading those results to make
affordable (from a UX performance point of view) more complex
chains or processes for subset of the results that occur later in the
overall answer.