Smooth scrolling and fix partial citiation
This smooths out the rate streamed content is presented to the user.
It enables a more complex backend to provide intial results quickly
and then use the fact the user will be reading those results to make
affordable (from a UX performance point of view) more complex
chains or processes for subset of the results that occur later in the
overall answer.

Prevent partial citation text

Citations are provided as text within square brackets.
Printing these before the citation is closed causes flashing during the
streaming.

This update checks if the text ends on something looks like a citation
in progress and does not print that text until the citation is closed.

When streaming is done we display with the non-streaming ui. This
means the risk of accidentally omitting text due to an unexpected
use of mismatched brackets by the llm will merely delay rather than
omit text. (Note .. any unexpected uses of brackets probably interacts
poorly with the citations, but I have not dug further on that.)