feat: Standardise LLM prompts with Prompty templates (#20)
Introduce templates to standardise LLM prompts and support switching between different models at runtime. Added a templates directory to store structured templates for ChatGPT 3.5 and MistralAI-Instruct (from Hugging Face).

Each template consists of the configuration for the model parameters written in YAML, and a Jinja template for prompting models. The backend for Ask and Chat tabs has been modified to use the Prompty objects available in the Promptflow package to load the templates. This feature addresses the issue of incompatible prompting for OpenAI services and the Hugging Face inference API through universalised, structured templates.

Signed-off-by: Alice Cui <alice7301@gmail.com>
Signed-off-by: Raimund Matulevic <rajmukx22@gmail.com>
Reviewed-by: Phoevos Kalemkeris <phoevos@protonmail.com>